Fixed rest of bugs for #17
